Add methods to EST_Wave for saving wave files by parts
Offering to write either only the file header or the actual data.

Additionally it provides a function to write a wave file header without
having an EST_Wave instance.

This patch is ground work for festival improvements in text2wave
memory (and disk) usage on long texts.

Backwards compatibility is preserved.