Cosmetic change makes builds fully reproducible
The `date` command used in these two scripts may have
some minor convenience to users, but it prevents the
build to be 100% reproducible.

It's just a cosmetic issue on two text files, but the
ReproducibleBuilds team in Debian [1] devoted the time to
prepare a patch [2] just to workaround this date issue.

[1] https://wiki.debian.org/ReproducibleBuilds/About
[2] https://salsa.debian.org/tts-team/speech-tools/-/blob/master/debian/patches/timestamp